feat: basic address book support
Only a hard-coded address book is supported at the moment. This commit
brings address book support to the `call` command only but follow-up
commits will expand support to other commands as well.

Address book support allows users to specify contract addresses and
calldata as `addr:contract-name`. For example, to check the ETH balance
of the zkLend smart contract:

    starkli call addr:eth balanceOf addr:zklend

Taking the `addr:` here as the "scheme", we can support other ways to
resolve arguments in the future by adding more schemes.