feat: work around sierra abi deser issues
The Sierra-CASM compiler 2.0.0 can handle 1.1.0 Sierra bytecode just
fine. However, it's unable to deserialize the an entire Sierra class
from 1.1.0 due to breaking changes in the ABI. The compiler does not
need the ABI: it's just part of the deserialization process. Therefore,
it's safe to simply remove it from the compiler input.