feat: auto check if class has been declared
`starkli` no longer attempts a class declaration when it's already
declared. It exits early with exit code `0` as it's not treated as an
error. Note that if an attempt is made and still fails due to a race
condition, `starkli` would consider it a failure and still exits with
error like it did before.