feat: parse fetched classes into original artifacts
Adds a new `--parse` flag to the `class-at` and `class-by-hash`
commands. When used, Starkli attempts to parse the raw JSON-RPC class
response back into the original contract artifact format (same as the
output from a compiler).

With this flag, users can retrieve the full original contract artifact
solely from a contract address or class hash. This is helpful when
trying to re-declare a certain class into another network.