feat: allow chain ID duplications in profiles
There are legitimate use cases for defining networks with the same chain
IDs - devnets, different RPC URLs for accessing the same network, etc..
Starkli should not enforce chain ID uniqueness.

One unfortunate side effect of this is that when using `--rpc` directly,
we can no longer uniquely match the chain ID fetched to a network
defined in the profile (this matching doesn't matter now, but will do in
the future when we have network-specific settings). This is not ideal
but still fine as it's expected that those who define custom networks
are likely to use `--network` instead of `--rpc` anyways. And by using
`--network` directly, users can always uniquely choose the network they
want the settings applied.