fix: tx submission error details not displayed
Error details are correctly printed when calling functions or estimating
fees. However, only the error type is printed when an error happens on
transaction submission. This is happening rather frequently recently
due to the increase of class declaration fees, as many users do not have
sufficient funds to cover the `max_fee` setting, causing validation
errors on transaction submission.